src/app/core/components/modal-confirmacion/modal-confirmacion.component.ts
import { Component, Inject } from '@angular/core';
import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';

export interface ModalModel {
  titulo: string;
  mensaje: string;
}
@Component({
  selector: 'app-modal-confirmacion',
  templateUrl: './modal-confirmacion.component.html',
  styleUrls: ['./modal-confirmacion.component.scss'],
})
export class ModalConfirmacionComponent {
  titulo: string;
  mensaje: string;

  constructor(
    public dialogRef: MatDialogRef<ModalConfirmacionComponent>,
    @Inject(MAT_DIALOG_DATA) public data: ModalModel,
  ) {
    this.titulo = data.titulo || 'Confirmación';
    this.mensaje = data.mensaje || '¿Estás seguro de realizar esta acción?';
  }

  cancelar(): void {
    this.dialogRef.close(false);
  }

  confirmar(): void {
    this.dialogRef.close(true);
  }
}
